requirements.txt
# Aplicación principal
streamlit>=1.28.0

# Procesamiento de PDFs
PyPDF2>=3.0.1

# Modelos de embeddings y ML
sentence-transformers>=2.2.2
torch>=2.0.0

# Análisis numérico y científico
numpy>=1.24.0
scikit-learn>=1.3.0

# Manipulación de datos
pandas>=2.0.0

# Utilidades adicionales
python-dotenv>=1.0.0
requests>=2.31.0

# Para desarrollo y testing (opcional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0